MGMT-15963: Guarantee that dry inventory will return a physical device
Presently, the dey inventory config will return the first found interface.
Often this ends up being the loopback device.

Assisted does not consider this to be part of inventory, therefore an invalid inventory is generated.

This PR aims to fix that by returning inventory for the first physical device in this case.